@@ -16,27 +16,27 @@ publish_to: none
 version: 1.9.33+286
 
 environment:
-  sdk: ">=3.0.0"
+  sdk: ">=3.1.4"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  dio: ^5.1.2
+  dio: ^5.3.4
   crypto: ^3.0.2
   path_provider: ^2.1.1
   shared_preferences: ^2.2.1
   share_plus: ^7.1.0
   permission_handler: ^11.0.0
-  archive: ^3.4.4
+  archive: ^3.4.9
   url_launcher: ^6.1.14
   bot_toast: ^4.0.4
-  package_info_plus: ^4.1.0
+  package_info_plus: ^5.0.1
   mobx: ^2.1.4
   flutter_mobx: ^2.0.6+5
   in_app_purchase: ^3.1.7
-  flutter_widget_from_html: ^0.10.3
+  flutter_widget_from_html: ^0.14.7
   device_info_plus: ^9.0.3
   flutter_displaymode: ^0.6.0
   receive_sharing_intent: ^1.4.5
@@ -55,20 +55,20 @@ dependencies:
   flutter_inappwebview: ^5.8.0
   auto_animated: ^3.1.0
   http: ^0.13.4
-  dio_cache_interceptor: ^3.4.2
+  dio_cache_interceptor: ^3.4.4
   photo_view: ^0.14.0
   http_interceptor: 2.0.0-beta.2
-  image: ^4.0.17
-  easy_refresh: ^3.3.2+1
-  flutter_hooks: ^0.18.6
-  hooks_riverpod: ^2.3.6
-  sqflite: ^2.2.8+4
-  dynamic_color: ^1.6.5
+  image: ^4.1.3
+  easy_refresh: ^3.3.2+4
+  flutter_hooks: ^0.20.3
+  hooks_riverpod: ^2.4.8
+  sqflite: ^2.3.0
+  dynamic_color: ^1.6.8
   # Design Style
-  fluent_ui: ^4.7.5
+  fluent_ui: ^4.8.0
   # Desktop Support
   flutter_acrylic: ^1.1.3
-  window_manager: ^0.3.5
+  window_manager: ^0.3.7
 
 dev_dependencies:
   flutter_test:
