@@ -111,63 +111,68 @@ class _NewIllustPageState extends State<NewIllustPage> {
   }
 
   Container buildContainer(BuildContext context) {
+    FlyoutController controller = FlyoutController();
     return Container(
-        child: Align(
-      alignment: Alignment.centerRight,
-      child: IconButton(
-          icon: Icon(FluentIcons.list),
-          onPressed: () {
-            showBottomSheet(
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.vertical(
-                    top: Radius.circular(16),
+      child: Align(
+        alignment: Alignment.centerRight,
+        child: FlyoutTarget(
+          controller: controller,
+          child: IconButton(
+            icon: Icon(FluentIcons.list),
+            onPressed: () => controller.showFlyout(
+              builder: (context) => MenuFlyout(
+                items: [
+                  MenuFlyoutItem(
+                    text: Text(I18n.of(context).all),
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                      setState(() {
+                        futureGet = ApiForceSource(
+                          futureGet: (e) => apiClient.getFollowIllusts(
+                            'all',
+                            force: e,
+                          ),
+                          glanceKey: "follow_illust",
+                        );
+                      });
+                    },
                   ),
-                ),
-                context: context,
-                builder: (context) => SafeArea(
-                      child: Column(
-                        mainAxisSize: MainAxisSize.min,
-                        children: <Widget>[
-                          ListTile(
-                            title: Text(I18n.of(context).all),
-                            onPressed: () {
-                              Navigator.of(context).pop();
-                              setState(() {
-                                futureGet = ApiForceSource(
-                                    futureGet: (e) => apiClient
-                                        .getFollowIllusts('all', force: e),
-                                    glanceKey: "follow_illust");
-                              });
-                            },
+                  MenuFlyoutItem(
+                    text: Text(I18n.of(context).public),
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                      setState(() {
+                        futureGet = ApiForceSource(
+                          futureGet: (e) => apiClient.getFollowIllusts(
+                            'public',
+                            force: e,
                           ),
-                          ListTile(
-                            title: Text(I18n.of(context).public),
-                            onPressed: () {
-                              Navigator.of(context).pop();
-                              setState(() {
-                                futureGet = ApiForceSource(
-                                    futureGet: (e) => apiClient
-                                        .getFollowIllusts('public', force: e),
-                                    glanceKey: "follow_illust");
-                              });
-                            },
+                          glanceKey: "follow_illust",
+                        );
+                      });
+                    },
+                  ),
+                  MenuFlyoutItem(
+                    text: Text(I18n.of(context).private),
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                      setState(() {
+                        futureGet = ApiForceSource(
+                          futureGet: (e) => apiClient.getFollowIllusts(
+                            'private',
+                            force: e,
                           ),
-                          ListTile(
-                            title: Text(I18n.of(context).private),
-                            onPressed: () {
-                              Navigator.of(context).pop();
-                              setState(() {
-                                futureGet = ApiForceSource(
-                                    futureGet: (e) => apiClient
-                                        .getFollowIllusts('private', force: e),
-                                    glanceKey: "follow_illust");
-                              });
-                            },
-                          ),
-                        ],
-                      ),
-                    ));
-          }),
-    ));
+                          glanceKey: "follow_illust",
+                        );
+                      });
+                    },
+                  ),
+                ],
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
   }
 }
