@@ -21,11 +21,11 @@ import 'dart:math';
 import 'package:easy_refresh/easy_refresh.dart';
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:pixez/component/pixez_default_header.dart';
 import 'package:pixez/er/leader.dart';
 import 'package:pixez/exts.dart';
 import 'package:pixez/fluent/component/illust_card.dart';
 import 'package:pixez/fluent/component/pixez_button.dart';
-import 'package:pixez/fluent/component/pixez_default_header.dart';
 import 'package:pixez/fluent/component/pixiv_image.dart';
 import 'package:pixez/fluent/page/hello/recom/recom_user_road.dart';
 import 'package:pixez/fluent/page/soup/soup_page.dart';
