@@ -17,6 +17,7 @@
 import 'package:easy_refresh/easy_refresh.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:pixez/component/pixez_default_header.dart';
 import 'package:pixez/component/pixiv_image.dart';
 import 'package:pixez/i18n.dart';
 import 'package:pixez/main.dart';
@@ -63,9 +64,10 @@ class _NovelUserBookmarkPageState extends State<NovelUserBookmarkPage> {
             onRefresh: () {
               _store.fetch();
             },
-            header: ClassicHeader(
+            header: PixezDefault.header(context,
                 position: IndicatorPosition.locator, safeArea: false),
-            footer: ClassicFooter(
+            footer: PixezDefault.footer(
+              context,
               position: IndicatorPosition.locator,
             ),
             childBuilder: (_, phy) {
@@ -185,8 +187,9 @@ class _NovelUserBookmarkPageState extends State<NovelUserBookmarkPage> {
                                           horizontal: 1),
                                       child: Text(
                                         f.name,
-                                        style:
-                                            Theme.of(context).textTheme.bodySmall,
+                                        style: Theme.of(context)
+                                            .textTheme
+                                            .bodySmall,
                                       ),
                                     )
                                 ],
